requirements.txt
websockets
python-dotenv
rapidfuzz
yt-dlp
